extract json file dependent code for modifying layouts
