+++ QualityControl/lib/services/JsonFileService.js
@@ -12,10 +12,8 @@
  * or submit itself to any jurisdiction.
  */
 
-import { LogManager, NotFoundError } from '@aliceo2/web-ui';
-const logger = LogManager.getLogger(`${process.env.npm_config_log_label ?? 'qcg'}/json`);
-import fs from 'fs';
-import path from 'path';
+import { NotFoundError } from '@aliceo2/web-ui';
+import LayoutRepository from '../database/repositories/LayoutRepository';
 
 /**
  * Store layouts inside JSON based file with atomic write
@@ -26,94 +24,7 @@ export class JsonFileService {
    * @param {string} pathname - path to JSON DB file
    */
   constructor(pathname) {
-    // Path of the file to store data
-    this.pathname = path.join(pathname);
-
-    // Path for writing file
-    this.pathnameTmp = `${this.pathname}~tmp`;
-
-    // Mirror data from content of JSON file
-    this.data = { layouts: [], users: [] };
-
-    // Write lock access
-    this.lock = new Lock();
-
-    this._syncFileAndInternalState();
-  }
-
-  /**
-   * Synchronize DB file content and `this.data` property
-   * @returns {undefined}
-   */
-  async _syncFileAndInternalState() {
-    await this._readFromFile();
-    await this._writeToFile();
-    logger.info(`Preferences will be saved in ${this.pathname}`);
-  }
-
-  /**
-   * Method to read from file and update the data variable
-   * @returns {Promise<undefined.Error>} - rejects if unable to read file
-   */
-  async _readFromFile() {
-    return new Promise((resolve, reject) => {
-      fs.readFile(this.pathname, (err, data) => {
-        if (err) {
-          // File does not exist, it's ok, we will create it
-          if (err.code === 'ENOENT') {
-            logger.info('DB file does not exist, will create one');
-            return resolve();
-          }
-
-          // Other errors reading
-          return reject(err);
-        }
-
-        try {
-          const dataFromFile = JSON.parse(data);
-
-          // Check data we just read
-          if (!dataFromFile || !dataFromFile.layouts || !Array.isArray(dataFromFile.layouts)) {
-            return reject(new Error(`DB file should have an array of layouts ${this.pathname}`));
-          }
-          // Check if users exists and if not declare and initialize with an empty array
-          if (!dataFromFile.users || !Array.isArray(dataFromFile.users)) {
-            dataFromFile.users = [];
-          }
-          this.data = dataFromFile;
-          resolve();
-        } catch {
-          return reject(new Error(`Unable to parse DB file ${this.pathname}`));
-        }
-      });
-    });
-  }
-
-  /**
-   * Write data to disk, atomically, with lock
-   * @returns {undefined}
-   */
-  async _writeToFile() {
-    await this.lock.acquire();
-
-    await new Promise((resolve, reject) => {
-      const dataToFile = JSON.stringify(this.data, null, 1);
-
-      fs.writeFile(this.pathnameTmp, dataToFile, (err) => {
-        if (err) {
-          return reject(err);
-        }
-        fs.rename(this.pathnameTmp, this.pathname, (err) => {
-          if (err) {
-            return reject(err);
-          }
-          logger.info('DB file updated');
-          resolve();
-        });
-      });
-    });
-
-    this.lock.release();
+    this.layoutRepository = new LayoutRepository(pathname);
   }
 
   /**
@@ -129,12 +40,11 @@ export class JsonFileService {
       throw new Error('layout name is mandatory');
     }
 
-    const layout = this.data.layouts.find((layout) => layout.id === newLayout.id);
+    const layout = this.layoutRepository.findLayoutById(newLayout.id);
     if (layout) {
       throw new Error(`layout with this id (${layout.id}) already exists`);
     }
-    this.data.layouts.push(newLayout);
-    await this._writeToFile();
+    this.layoutRepository.createLayout(newLayout);
     return newLayout;
   }
 
@@ -144,8 +54,8 @@ export class JsonFileService {
    * @returns {Layout} - layout object
    * @throws {Error}
    */
-  async readLayout(layoutId) {
-    const layout = this.data.layouts.find((layout) => layout.id === layoutId);
+  readLayout(layoutId) {
+    const layout = this.layoutRepository.findLayoutById(layoutId);
     if (!layout) {
       throw new NotFoundError(`layout (${layoutId}) not found`);
     }
@@ -158,8 +68,8 @@ export class JsonFileService {
    * @returns {Layout} - object with layout information
    * @throws
    */
-  async readLayoutByName(layoutName) {
-    const layout = this.data.layouts.find((layout) => layout.name === layoutName);
+  readLayoutByName(layoutName) {
+    const layout = this.layoutRepository.findLayoutByName(layoutName);
     if (!layout) {
       throw new NotFoundError(`Layout (${layoutName}) not found`);
     }
@@ -172,10 +82,9 @@ export class JsonFileService {
    * @param {Layout} data - layout new data
    * @returns {object} Empty details
    */
-  async updateLayout(layoutId, data) {
-    const layout = await this.readLayout(layoutId);
-    Object.assign(layout, data);
-    this._writeToFile();
+  updateLayout(layoutId, data) {
+    const layout = this.readLayout(layoutId);
+    this.layoutRepository.updateLayout(layout, data);
     return layoutId;
   }
 
@@ -184,11 +93,9 @@ export class JsonFileService {
    * @param {string} layoutId - id of the layout to be removed
    * @returns {object} Empty details
    */
-  async deleteLayout(layoutId) {
-    const layout = await this.readLayout(layoutId);
-    const index = this.data.layouts.indexOf(layout);
-    this.data.layouts.splice(index, 1);
-    await this._writeToFile();
+  deleteLayout(layoutId) {
+    const layout = this.readLayout(layoutId);
+    this.layoutRepository.deleteLayout(layout);
     return layoutId;
   }
 
@@ -198,9 +105,7 @@ export class JsonFileService {
    * @returns {Array<Layout>} - list of layouts as per the filter
    */
   async listLayouts(filter = {}) {
-    return this.data.layouts.filter((layout) =>
-      (filter.owner_id === undefined || layout.owner_id === filter.owner_id)
-      && (filter.name === undefined || layout.name === filter.name));
+    return this.layoutRepository.listLayouts(filter);
   }
 
   /**
@@ -264,58 +169,5 @@ export class JsonFileService {
     if (isNaN(user.id)) {
       throw new Error('Field id must be a number');
     }
-  }
-}
-
-/**
- * Simple Lock blocked Promise for exclusive access to resource
- * @example
- * let lock = new Lock();
- * lock.acquire();
- * setTimeout(() => lock.release(), 1000);
- * await lock.acquire(); // will wait 1000ms
- */
-class Lock {
-  /**
-   * Initialize lock to released
-   */
-  constructor() {
-    this._locked = false;
-    this._queue = []; // Callbacks of next owners of the lock
-  }
-
-  /**
-   * Acquires lock if available and returns immediately
-   * otherwise wait for lock to be released
-   * @returns {Promise} - result of the lock
-   */
-  acquire() {
-    return new Promise((resolve) => {
-      // If nobody has the lock, take it and resolve immediately
-      if (!this._locked) {
-        this._locked = true;
-        return resolve();
-      }
-
-      // Otherwise, push as next owner
-      this._queue.push(resolve);
-    });
-  }
-
-  /**
-   * Releases lock and give it to next in queue if any
-   * @returns {object|undefined} - next owner of the lock
-   */
-  release() {
-    // Release the lock immediately
-    setImmediate(() => {
-      const nextOwner = this._queue.shift();
-      if (nextOwner) {
-        this._locked = true;
-        return nextOwner();
-      }
-
-      this._locked = false;
-    });
   }
 }
